去除初始流量的整数限制

+++ src/Controllers/AuthController.php
@@ -393,7 +393,7 @@ class AuthController extends BaseController
         $user->im_type              = $imtype;
         $user->im_value             = $antiXss->xss_clean($imvalue);
         
-        $user->transfer_enable      = Tools::toGB((int) Config::getconfig('Register.string.defaultTraffic'));
+        $user->transfer_enable      = Tools::toGB(Config::getconfig('Register.string.defaultTraffic'));
         $user->invite_num           = (int) Config::getconfig('Register.string.defaultInviteNum');
         $user->auto_reset_day       = $_ENV['reg_auto_reset_day'];
         $user->auto_reset_bandwidth = $_ENV['reg_auto_reset_bandwidth'];
